Rename EBLEnvelope{,TO} to EndorsementChainEntry{,TO}
Perform the rename every where except if it would affect the external
API. The omission ensures that we can use existing CI to verify the
rename did not break anything and keep the code working (on the
previous version of the API).

Also added a couple of TODOs when I spotted something wrong. However,
I did not fix them in this commit as it would mix refactoring with
semantic changes and would make this commit considerably harder to
review.

Signed-off-by: Niels Thykier <nit@globeteam.com>